+++ @angular-generic-table/core/index.ts
@@ -12,3 +12,4 @@ export {GtRow} from './interfaces/gt-row';
 export {GtTexts} from './interfaces/gt-texts';
 export {GtOptions} from './interfaces/gt-options';
 export {GtCustomComponent} from './components/gt-custom-component-factory';
+export {GtHighlightPipe} from './pipes/gt-highlight.pipe';